feat: add PDG integration to dashboard
