+++ test/0.8.25/vaults/delegation/delegation.test.ts
@@ -115,6 +115,7 @@ describe("Delegation.sol", () => {
         exitRequester,
         disconnecter,
         curator,
+        assetRecoverer: curator,
         nodeOperatorManager,
         nodeOperatorFeeClaimer,
         curatorFeeBP: 0n,
